[DOCS] Document MacOS keymap more prominently
It is very confusing to new Gertty users like me that it seems to
function fine except the search key is a noop on MacOS, and the
current documentation for fixing this is buried in a 250 line
configuration example.

This is just a minor documentation fix to make the possible need
for a custom key mapping on MacOS more prominent.

Change-Id: I8053de3a0a64ebeeec9d2d400766a8debe5c17fa