Add a Dockerfile for docker to docker
This offers a slightly higher degree of isolation than virtualenv, so
users don't even need python or pip or virtualenv installed.

Adding another option highlighted that the docs are confusing as they
rapidly transition from the source install methods to configuration, so
we add a header to make it more clear where our change ends.

Change-Id: Ia072c5f8d39273285dd2ccc7b611ba893e2e7c0b