gitrepo DiffFile: convert tab to » + spaces
tabs are displayed as question mark by urwid, so need replacing.
While we are replacing, might as well mark them like in the web UI.

Ideally I would want to use a slightly different color for them,
but that would require to add many new colors in the config (context,
full new line, part of new line) and would not be very compatible
with 8-colors terminals -- the '»' character is hopefully rarely used
enough to be useable as is.

Change-Id: Ic974774bf546765fa5b3e9266aee8ff973865502