Fix tab expansion in inter-patchset diff
There are some lines which were being read as byte arrays and
were not decoded into strings which was causing them to bypass
tab expansion (because they caused an exception which was caught
and the un-expanded text was used instead).  Correct those cases
and disable the exception handler so that we can identify any
other cases by triggering a crash.

Change-Id: I4b34132290fc8611cc770d572068197b33c0789c