Fix change view display of inline comments in python3
File comments were causing an exception when displayed on the
change view because python3 does not allow comparing None to int,
and file comments have a line number of None.

To correct this, use 0 as the line number for file comments whin
sorting and displaying them in the change view; there is no line
0 so this minor aliasing shouldn't cause any collisions.

Also, add a newline to the end of file comments, just as regular
inline comments (otherwise if a file comment was followed by a
line comment, they would appear on the same line).

Change-Id: Iedf88cdfd83adbc405e26cb013c9f57481d50c90